Refactored XML loading
Took game building responsiblities off the custom SAX parser handler (now called ArtGameSaveFileHandler) by creating class GameMaker, which is responsible for the game building instead. To avoid type casting (and to get extra marks by showing off design patterns) the object builders use the strategy patern, taking a build strategy as a constructor argument. There is a unique build strategy for each type of object in the game. Also, ObjectBuilders no longer have the method buildObject(). Instead the build strategies typically take the GameMaker as a constructor argument. When the method addToGame() is called on an ObjectBuilder, the build strategy calls the appropriate methods on GameMaker for adding it's type of object