Can now save Patrol paths
Uses step element for each coordinate of guards path instead of stretches. Also changed the painting list in ArtGameSaver to painting set to avoid writing the same art element twice. @kaishuoy Noticed bug where if a painting in on two walls and a player takes it off one wall, the painting is still on the other wall, as well as in the player's inventory